+++ szdiagnostic/szdiagnostic.go
@@ -39,7 +39,7 @@ const (
 // ----------------------------------------------------------------------------
 
 /*
-Method CheckDatastorePerformance runs performance tests on the Senzing datastore.
+Method CheckRepositoryPerformance runs performance tests on the Senzing repository.
 
 Input
   - ctx: A context to control lifecycle.
@@ -50,7 +50,7 @@ Output
   - A JSON document containing performance results.
     Example: `{"numRecordsInserted":0,"insertTime":0}`
 */
-func (client *Szdiagnostic) CheckDatastorePerformance(ctx context.Context, secondsToRun int) (string, error) {
+func (client *Szdiagnostic) CheckRepositoryPerformance(ctx context.Context, secondsToRun int) (string, error) {
 	var (
 		err    error
 		result string
@@ -60,10 +60,11 @@ func (client *Szdiagnostic) CheckDatastorePerformance(ctx context.Context, secon
 		client.traceEntry(1, secondsToRun)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(2, secondsToRun, result, err, time.Since(entryTime)) }()
 	}
 
-	result, err = client.checkDatastorePerformance(ctx, secondsToRun)
+	result, err = client.checkRepositoryPerformance(ctx, secondsToRun)
 
 	if client.observers != nil {
 		go func() {
@@ -76,16 +77,16 @@ func (client *Szdiagnostic) CheckDatastorePerformance(ctx context.Context, secon
 }
 
 /*
-Method GetDatastoreInfo returns information about the Senzing datastore.
+Method GetRepositoryInfo returns information about the Senzing repository.
 
 Input
   - ctx: A context to control lifecycle.
 
 Output
 
-  - A JSON document containing Senzing datastore metadata.
+  - A JSON document containing Senzing repository metadata.
 */
-func (client *Szdiagnostic) GetDatastoreInfo(ctx context.Context) (string, error) {
+func (client *Szdiagnostic) GetRepositoryInfo(ctx context.Context) (string, error) {
 	var (
 		err    error
 		result string
@@ -95,10 +96,11 @@ func (client *Szdiagnostic) GetDatastoreInfo(ctx context.Context) (string, error
 		client.traceEntry(7)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(8, result, err, time.Since(entryTime)) }()
 	}
 
-	result, err = client.getDatastoreInfo(ctx)
+	result, err = client.getRepositoryInfo(ctx)
 
 	if client.observers != nil {
 		go func() {
@@ -132,6 +134,7 @@ func (client *Szdiagnostic) GetFeature(ctx context.Context, featureID int64) (st
 		client.traceEntry(9, featureID)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(10, featureID, result, err, time.Since(entryTime)) }()
 	}
 
@@ -150,7 +153,7 @@ func (client *Szdiagnostic) GetFeature(ctx context.Context, featureID int64) (st
 }
 
 /*
-WARNING: Method PurgeRepository removes every record in the Senzing datastore.
+WARNING: Method PurgeRepository removes every record in the Senzing repository.
 This is a destructive method that cannot be undone.
 Before calling purgeRepository(), all programs using Senzing MUST be terminated.
 
@@ -164,6 +167,7 @@ func (client *Szdiagnostic) PurgeRepository(ctx context.Context) error {
 		client.traceEntry(17)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(18, err, time.Since(entryTime)) }()
 	}
 
@@ -196,6 +200,7 @@ func (client *Szdiagnostic) Destroy(ctx context.Context) error {
 		client.traceEntry(5)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(6, err, time.Since(entryTime)) }()
 	}
 
@@ -247,6 +252,7 @@ func (client *Szdiagnostic) Initialize(
 		client.traceEntry(15, instanceName, settings, configID, verboseLogging)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(16, instanceName, settings, configID, verboseLogging, err, time.Since(entryTime))
 		}()
@@ -281,6 +287,7 @@ func (client *Szdiagnostic) RegisterObserver(ctx context.Context, observer obser
 		client.traceEntry(703, observer.GetObserverID(ctx))
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(704, observer.GetObserverID(ctx), err, time.Since(entryTime)) }()
 	}
 
@@ -316,6 +323,7 @@ func (client *Szdiagnostic) Reinitialize(ctx context.Context, configID int64) er
 		client.traceEntry(19, configID)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(20, configID, err, time.Since(entryTime)) }()
 	}
 
@@ -347,6 +355,7 @@ func (client *Szdiagnostic) SetLogLevel(ctx context.Context, logLevelName string
 		client.traceEntry(705, logLevelName)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(706, logLevelName, err, time.Since(entryTime)) }()
 	}
 
@@ -395,6 +404,7 @@ func (client *Szdiagnostic) UnregisterObserver(ctx context.Context, observer obs
 		client.traceEntry(707, observer.GetObserverID(ctx))
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(708, observer.GetObserverID(ctx), err, time.Since(entryTime)) }()
 	}
 
@@ -422,23 +432,23 @@ func (client *Szdiagnostic) UnregisterObserver(ctx context.Context, observer obs
 // Private methods for gRPC request/response
 // ----------------------------------------------------------------------------
 
-func (client *Szdiagnostic) checkDatastorePerformance(ctx context.Context, secondsToRun int) (string, error) {
+func (client *Szdiagnostic) checkRepositoryPerformance(ctx context.Context, secondsToRun int) (string, error) {
 	var result string
 
-	request := &szpb.CheckDatastorePerformanceRequest{
+	request := &szpb.CheckRepositoryPerformanceRequest{
 		SecondsToRun: int32(secondsToRun), //nolint:gosec
 	}
-	response, err := client.GrpcClient.CheckDatastorePerformance(ctx, request)
+	response, err := client.GrpcClient.CheckRepositoryPerformance(ctx, request)
 	result = response.GetResult()
 
 	return result, helper.ConvertGrpcError(err)
 }
 
-func (client *Szdiagnostic) getDatastoreInfo(ctx context.Context) (string, error) {
+func (client *Szdiagnostic) getRepositoryInfo(ctx context.Context) (string, error) {
 	var result string
 
-	request := &szpb.GetDatastoreInfoRequest{}
-	response, err := client.GrpcClient.GetDatastoreInfo(ctx, request)
+	request := &szpb.GetRepositoryInfoRequest{}
+	response, err := client.GrpcClient.GetRepositoryInfo(ctx, request)
 	result = response.GetResult()
 
 	return result, helper.ConvertGrpcError(err)