+++ szdiagnostic/szdiagnostic_examples_test.go
@@ -16,7 +16,7 @@ const AllLines = -1
 // Interface methods - Examples for godoc documentation
 // ----------------------------------------------------------------------------
 
-func ExampleSzdiagnostic_CheckDatastorePerformance() {
+func ExampleSzdiagnostic_CheckRepositoryPerformance() {
 	// For more information, visit
 	// https://github.com/senzing-garage/sz-sdk-go-grpc/blob/main/szdiagnostic/szdiagnostic_examples_test.go
 	ctx := context.TODO()
@@ -29,7 +29,7 @@ func ExampleSzdiagnostic_CheckDatastorePerformance() {
 
 	secondsToRun := 1
 
-	result, err := szDiagnostic.CheckDatastorePerformance(ctx, secondsToRun)
+	result, err := szDiagnostic.CheckRepositoryPerformance(ctx, secondsToRun)
 	if err != nil {
 		handleError(err)
 	}
@@ -42,7 +42,7 @@ func ExampleSzdiagnostic_CheckDatastorePerformance() {
 	// }
 }
 
-func ExampleSzdiagnostic_GetDatastoreInfo() {
+func ExampleSzdiagnostic_GetRepositoryInfo() {
 	// For more information, visit
 	// https://github.com/senzing-garage/sz-sdk-go-grpc/blob/main/szdiagnostic/szdiagnostic_examples_test.go
 	ctx := context.TODO()
@@ -53,7 +53,7 @@ func ExampleSzdiagnostic_GetDatastoreInfo() {
 		handleError(err)
 	}
 
-	result, err := szDiagnostic.GetDatastoreInfo(ctx)
+	result, err := szDiagnostic.GetRepositoryInfo(ctx)
 	if err != nil {
 		handleError(err)
 	}