+++ szengine/szengine.go
@@ -41,10 +41,10 @@ const (
 // ----------------------------------------------------------------------------
 
 /*
-Method AddRecord adds a record into the Senzing datastore.
+Method AddRecord adds a record into the Senzing repository.
 The unique identifier of a record is the [dataSourceCode, recordID] compound key.
-If the unique identifier does not exist in the Senzing datastore, a new record definition is created in the
-Senzing datastore.
+If the unique identifier does not exist in the Senzing repository, a new record definition is created in the
+Senzing repository.
 If the unique identifier already exists, the new record definition will replace the old record definition.
 If the record definition contains JSON keys of `DATA_SOURCE` and/or `RECORD_ID`, they must match the values of `
 dataSourceCode` and `recordID`.
@@ -53,7 +53,7 @@ Input
   - ctx: A context to control lifecycle.
   - dataSourceCode: Identifies the provenance of the data.
   - recordID: The unique identifier within the records of the same data source.
-  - recordDefinition: A JSON document containing the record to be added to the Senzing datastore.
+  - recordDefinition: A JSON document containing the record to be added to the Senzing repository.
   - flags: Flags used to control information returned.
 
 Output
@@ -75,6 +75,7 @@ func (client *Szengine) AddRecord(
 		client.traceEntry(1, dataSourceCode, recordID, recordDefinition, flags)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(2, dataSourceCode, recordID, recordDefinition, flags, result, err, time.Since(entryTime))
 		}()
@@ -97,27 +98,29 @@ func (client *Szengine) AddRecord(
 }
 
 /*
-Method CloseExport closes the exported document created by [Szengine.ExportJSONEntityReport] or
+Method CloseExportReport closes the exported document created by [Szengine.ExportJSONEntityReport] or
 [Szengine.ExportCsvEntityReport].
-It is part of the ExportXxxEntityReport(), [Szengine.FetchNext], CloseExport lifecycle of a list of entities to export.
-CloseExport is idempotent; an exportHandle may be closed multiple times.
+It is part of the ExportXxxEntityReport(), [Szengine.FetchNext],
+CloseExportReport lifecycle of a list of entities to export.
+CloseExportReport is idempotent; an exportHandle may be closed multiple times.
 
 Input
   - ctx: A context to control lifecycle.
   - exportHandle: A handle created by [Szengine.ExportJSONEntityReport] or [Szengine.ExportCsvEntityReport]
     that is to be closed.
 */
-func (client *Szengine) CloseExport(ctx context.Context, exportHandle uintptr) error {
+func (client *Szengine) CloseExportReport(ctx context.Context, exportHandle uintptr) error {
 	var err error
 
 	if client.isTrace {
 		client.traceEntry(5, exportHandle)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(6, exportHandle, err, time.Since(entryTime)) }()
 	}
 
-	err = client.closeExport(ctx, exportHandle)
+	err = client.closeExportReport(ctx, exportHandle)
 
 	if client.observers != nil {
 		go func() {
@@ -149,6 +152,7 @@ func (client *Szengine) CountRedoRecords(ctx context.Context) (int64, error) {
 		client.traceEntry(7)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(8, result, err, time.Since(entryTime)) }()
 	}
 
@@ -165,11 +169,11 @@ func (client *Szengine) CountRedoRecords(ctx context.Context) (int64, error) {
 }
 
 /*
-Method DeleteRecord deletes a record from the Senzing datastore.
+Method DeleteRecord deletes a record from the Senzing repository.
 The unique identifier of a record is the [dataSourceCode, recordID] compound key.
 DeleteRecord() is idempotent.
 Multiple calls to delete the same unique identifier will all succeed,
-even if the unique identifier is not present in the Senzing datastore.
+even if the unique identifier is not present in the Senzing repository.
 
 Input
   - ctx: A context to control lifecycle.
@@ -195,6 +199,7 @@ func (client *Szengine) DeleteRecord(
 		client.traceEntry(9, dataSourceCode, recordID, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(10, dataSourceCode, recordID, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -216,7 +221,7 @@ func (client *Szengine) DeleteRecord(
 
 /*
 Method ExportCsvEntityReport initializes a cursor over a CSV document of exported entities.
-It is part of the ExportCsvEntityReport, [Szengine.FetchNext], [Szengine.CloseExport] lifecycle
+It is part of the ExportCsvEntityReport, [Szengine.FetchNext], [Szengine.CloseExportReport] lifecycle
 of a list of entities to export.
 The first exported line is the CSV header.
 Each subsequent line contains metadata for a single entity.
@@ -240,6 +245,7 @@ func (client *Szengine) ExportCsvEntityReport(ctx context.Context, csvColumnList
 		client.traceEntry(13, csvColumnList, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(14, csvColumnList, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -260,7 +266,7 @@ func (client *Szengine) ExportCsvEntityReport(ctx context.Context, csvColumnList
 /*
 Method ExportCsvEntityReportIterator creates an Iterator that can be used in a for-loop
 to scroll through a CSV document of exported entities.
-It is a convenience method for the [Szenzine.ExportCsvEntityReport], [Szengine.FetchNext], [Szengine.CloseExport]
+It is a convenience method for the [Szenzine.ExportCsvEntityReport], [Szengine.FetchNext], [Szengine.CloseExportReport]
 lifecycle of a list of entities to export.
 
 Input
@@ -278,6 +284,7 @@ func (client *Szengine) ExportCsvEntityReportIterator(
 	flags int64,
 ) chan senzing.StringFragment {
 	stringFragmentChannel := make(chan senzing.StringFragment)
+
 	go func() {
 		defer close(stringFragmentChannel)
 
@@ -287,6 +294,7 @@ func (client *Szengine) ExportCsvEntityReportIterator(
 			client.traceEntry(15, csvColumnList, flags)
 
 			entryTime := time.Now()
+
 			defer func() { client.traceExit(16, csvColumnList, flags, err, time.Since(entryTime)) }()
 		}
 
@@ -303,6 +311,7 @@ func (client *Szengine) ExportCsvEntityReportIterator(
 
 			return
 		}
+
 	forLoop:
 		for {
 			select {
@@ -318,12 +327,14 @@ func (client *Szengine) ExportCsvEntityReportIterator(
 					if errors.Is(err, io.EOF) {
 						break forLoop
 					}
+
 					stringFragmentChannel <- senzing.StringFragment{
 						Error: helper.ConvertGrpcError(err),
 					}
 
 					break forLoop
 				}
+
 				stringFragmentChannel <- senzing.StringFragment{
 					Value: response.GetResult(),
 				}
@@ -345,7 +356,7 @@ func (client *Szengine) ExportCsvEntityReportIterator(
 
 /*
 Method ExportJSONEntityReport initializes a cursor over a JSON document of exported entities.
-It is part of the ExportJSONEntityReport, [Szengine.FetchNext], [Szengine.CloseExport] lifecycle
+It is part of the ExportJSONEntityReport, [Szengine.FetchNext], [Szengine.CloseExportReport] lifecycle
 of a list of entities to export.
 
 Input
@@ -365,6 +376,7 @@ func (client *Szengine) ExportJSONEntityReport(ctx context.Context, flags int64)
 		client.traceEntry(17, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(18, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -385,7 +397,7 @@ func (client *Szengine) ExportJSONEntityReport(ctx context.Context, flags int64)
 /*
 Method ExportJSONEntityReportIterator creates an Iterator that can be used in a for-loop
 to scroll through a JSON document of exported entities.
-It is a convenience method for the [Szengine.ExportJSONEntityReport], [Szengine.FetchNext], [Szengine.CloseExport]
+It is a convenience method for the [Szengine.ExportJSONEntityReport], [Szengine.FetchNext], [Szengine.CloseExportReport]
 lifecycle of a list of entities to export.
 
 Input
@@ -397,6 +409,7 @@ Output
 */
 func (client *Szengine) ExportJSONEntityReportIterator(ctx context.Context, flags int64) chan senzing.StringFragment {
 	stringFragmentChannel := make(chan senzing.StringFragment)
+
 	go func() {
 		defer close(stringFragmentChannel)
 
@@ -406,6 +419,7 @@ func (client *Szengine) ExportJSONEntityReportIterator(ctx context.Context, flag
 			client.traceEntry(19, flags)
 
 			entryTime := time.Now()
+
 			defer func() { client.traceExit(20, flags, err, time.Since(entryTime)) }()
 		}
 
@@ -421,6 +435,7 @@ func (client *Szengine) ExportJSONEntityReportIterator(ctx context.Context, flag
 
 			return
 		}
+
 	forLoop:
 		for {
 			select {
@@ -436,12 +451,14 @@ func (client *Szengine) ExportJSONEntityReportIterator(ctx context.Context, flag
 					if errors.Is(err, io.EOF) {
 						break forLoop
 					}
+
 					stringFragmentChannel <- senzing.StringFragment{
 						Error: helper.ConvertGrpcError(err),
 					}
 
 					break forLoop
 				}
+
 				stringFragmentChannel <- senzing.StringFragment{
 					Value: response.GetResult(),
 				}
@@ -462,7 +479,7 @@ func (client *Szengine) ExportJSONEntityReportIterator(ctx context.Context, flag
 /*
 Method FetchNext is used to scroll through an exported JSON or CSV document.
 It is part of the [Szengine.ExportJSONEntityReport] or [Szengine.ExportCsvEntityReport], FetchNext,
-[Szengine.CloseExport] lifecycle of a list of exported entities.
+[Szengine.CloseExportReport] lifecycle of a list of exported entities.
 
 Input
   - ctx: A context to control lifecycle.
@@ -481,6 +498,7 @@ func (client *Szengine) FetchNext(ctx context.Context, exportHandle uintptr) (st
 		client.traceEntry(21, exportHandle)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(22, exportHandle, result, err, time.Since(entryTime)) }()
 	}
 
@@ -522,6 +540,7 @@ func (client *Szengine) FindInterestingEntitiesByEntityID(
 		client.traceEntry(23, entityID, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(24, entityID, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -568,6 +587,7 @@ func (client *Szengine) FindInterestingEntitiesByRecordID(
 		client.traceEntry(25, dataSourceCode, recordID, flags)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(26, dataSourceCode, recordID, flags, result, err, time.Since(entryTime))
 		}()
@@ -624,6 +644,7 @@ func (client *Szengine) FindNetworkByEntityID(
 		client.traceEntry(27, entityIDs, maxDegrees, buildOutDegrees, buildOutMaxEntities, flags)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(
 				28,
@@ -697,6 +718,7 @@ func (client *Szengine) FindNetworkByRecordID(
 		client.traceEntry(29, recordKeys, maxDegrees, buildOutDegrees, buildOutMaxEntities, flags)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(
 				30,
@@ -773,6 +795,7 @@ func (client *Szengine) FindPathByEntityID(
 		client.traceEntry(31, startEntityID, endEntityID, maxDegrees, avoidEntityIDs, requiredDataSources, flags)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(32, startEntityID, endEntityID, maxDegrees, avoidEntityIDs, requiredDataSources,
 				flags, result, err, time.Since(entryTime))
@@ -857,6 +880,7 @@ func (client *Szengine) FindPathByRecordID(
 			avoidRecordKeys, requiredDataSources, flags)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(34, startDataSourceCode, startRecordID, endDataSourceCode, endRecordID, maxDegrees,
 				avoidRecordKeys, requiredDataSources, flags, result, err, time.Since(entryTime))
@@ -912,6 +936,7 @@ func (client *Szengine) GetActiveConfigID(ctx context.Context) (int64, error) {
 		client.traceEntry(35)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(36, result, err, time.Since(entryTime)) }()
 	}
 
@@ -949,6 +974,7 @@ func (client *Szengine) GetEntityByEntityID(ctx context.Context, entityID int64,
 		client.traceEntry(37, entityID, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(38, entityID, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -995,6 +1021,7 @@ func (client *Szengine) GetEntityByRecordID(
 		client.traceEntry(39, dataSourceCode, recordID, flags)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(40, dataSourceCode, recordID, flags, result, err, time.Since(entryTime))
 		}()
@@ -1017,7 +1044,7 @@ func (client *Szengine) GetEntityByRecordID(
 }
 
 /*
-Method GetRecord returns a JSON document containing a single record from the Senzing datastore.
+Method GetRecord returns a JSON document containing a single record from the Senzing repository.
 
 Input
   - ctx: A context to control lifecycle.
@@ -1043,6 +1070,7 @@ func (client *Szengine) GetRecord(
 		client.traceEntry(45, dataSourceCode, recordID, flags)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(46, dataSourceCode, recordID, flags, result, err, time.Since(entryTime))
 		}()
@@ -1065,7 +1093,7 @@ func (client *Szengine) GetRecord(
 }
 
 /*
-Method GetRedoRecord returns the next maintenance record from the Senzing datastore.
+Method GetRedoRecord returns the next maintenance record from the Senzing repository.
 Usually, [Szengine.ProcessRedoRecord] is called to process the maintenance record retrieved by GetRedoRecord.
 
 Input
@@ -1084,6 +1112,7 @@ func (client *Szengine) GetRedoRecord(ctx context.Context) (string, error) {
 		client.traceEntry(47)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(48, result, err, time.Since(entryTime)) }()
 	}
 
@@ -1119,6 +1148,7 @@ func (client *Szengine) GetStats(ctx context.Context) (string, error) {
 		client.traceEntry(49)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(50, result, err, time.Since(entryTime)) }()
 	}
 
@@ -1160,6 +1190,7 @@ func (client *Szengine) GetVirtualEntityByRecordID(
 		client.traceEntry(51, recordKeys, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(52, recordKeys, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -1199,6 +1230,7 @@ func (client *Szengine) HowEntityByEntityID(ctx context.Context, entityID int64,
 		client.traceEntry(53, entityID, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(54, entityID, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -1218,17 +1250,17 @@ func (client *Szengine) HowEntityByEntityID(ctx context.Context, entityID int64,
 }
 
 /*
-Method PreprocessRecord tests adding a record into the Senzing datastore.
+Method GetRecordPreview tests adding a record into the Senzing repository.
 
 Input
   - ctx: A context to control lifecycle.
-  - recordDefinition: A JSON document containing the record to be tested against the Senzing datastore.
+  - recordDefinition: A JSON document containing the record to be tested against the Senzing repository.
   - flags: Flags used to control information returned.
 
 Output
   - A JSON document containing metadata as specified by the flags.
 */
-func (client *Szengine) PreprocessRecord(ctx context.Context, recordDefinition string, flags int64) (string, error) {
+func (client *Szengine) GetRecordPreview(ctx context.Context, recordDefinition string, flags int64) (string, error) {
 	var (
 		err    error
 		result string
@@ -1238,12 +1270,13 @@ func (client *Szengine) PreprocessRecord(ctx context.Context, recordDefinition s
 		client.traceEntry(77, recordDefinition, flags)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(78, recordDefinition, flags, result, err, time.Since(entryTime))
 		}()
 	}
 
-	result, err = client.preprocessRecord(ctx, recordDefinition, flags)
+	result, err = client.getRecordPreview(ctx, recordDefinition, flags)
 
 	if client.observers != nil {
 		go func() {
@@ -1270,6 +1303,7 @@ func (client *Szengine) PrimeEngine(ctx context.Context) error {
 		client.traceEntry(57)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(58, err, time.Since(entryTime)) }()
 	}
 
@@ -1305,6 +1339,7 @@ func (client *Szengine) ProcessRedoRecord(ctx context.Context, redoRecord string
 		client.traceEntry(59, redoRecord, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(60, redoRecord, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -1343,6 +1378,7 @@ func (client *Szengine) ReevaluateEntity(ctx context.Context, entityID int64, fl
 		client.traceEntry(61, entityID, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(62, entityID, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -1388,6 +1424,7 @@ func (client *Szengine) ReevaluateRecord(
 		client.traceEntry(63, dataSourceCode, recordID, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(64, dataSourceCode, recordID, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -1438,6 +1475,7 @@ func (client *Szengine) SearchByAttributes(
 		client.traceEntry(69, attributes, searchProfile, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(70, attributes, searchProfile, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -1484,6 +1522,7 @@ func (client *Szengine) WhyEntities(
 		client.traceEntry(71, entityID1, entityID2, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(72, entityID1, entityID2, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -1530,6 +1569,7 @@ func (client *Szengine) WhyRecordInEntity(
 		client.traceEntry(73, dataSourceCode, recordID, flags)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(74, dataSourceCode, recordID, flags, result, err, time.Since(entryTime)) }()
 	}
 
@@ -1580,6 +1620,7 @@ func (client *Szengine) WhyRecords(
 		client.traceEntry(75, dataSourceCode1, recordID1, dataSourceCode2, recordID2, flags)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(
 				76,
@@ -1645,6 +1686,7 @@ func (client *Szengine) WhySearch(
 		client.traceEntry(69, attributes, entityID, searchProfile, flags)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(70, attributes, entityID, searchProfile, flags, result, err, time.Since(entryTime))
 		}()
@@ -1684,6 +1726,7 @@ func (client *Szengine) Destroy(ctx context.Context) error {
 		client.traceEntry(11)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(12, err, time.Since(entryTime)) }()
 	}
 
@@ -1735,6 +1778,7 @@ func (client *Szengine) Initialize(
 		client.traceEntry(55, instanceName, settings, configID, verboseLogging)
 
 		entryTime := time.Now()
+
 		defer func() {
 			client.traceExit(56, instanceName, settings, configID, verboseLogging, err, time.Since(entryTime))
 		}()
@@ -1769,6 +1813,7 @@ func (client *Szengine) RegisterObserver(ctx context.Context, observer observer.
 		client.traceEntry(703, observer.GetObserverID(ctx))
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(704, observer.GetObserverID(ctx), err, time.Since(entryTime)) }()
 	}
 
@@ -1804,6 +1849,7 @@ func (client *Szengine) Reinitialize(ctx context.Context, configID int64) error 
 		client.traceEntry(65, configID)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(66, configID, err, time.Since(entryTime)) }()
 	}
 
@@ -1835,6 +1881,7 @@ func (client *Szengine) SetLogLevel(ctx context.Context, logLevelName string) er
 		client.traceEntry(705, logLevelName)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(706, logLevelName, err, time.Since(entryTime)) }()
 	}
 
@@ -1883,6 +1930,7 @@ func (client *Szengine) UnregisterObserver(ctx context.Context, observer observe
 		client.traceEntry(707, observer.GetObserverID(ctx))
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(708, observer.GetObserverID(ctx), err, time.Since(entryTime)) }()
 	}
 
@@ -1928,11 +1976,11 @@ func (client *Szengine) addRecord(
 	return result, helper.ConvertGrpcError(err)
 }
 
-func (client *Szengine) closeExport(ctx context.Context, exportHandle uintptr) error {
-	request := &szpb.CloseExportRequest{
+func (client *Szengine) closeExportReport(ctx context.Context, exportHandle uintptr) error {
+	request := &szpb.CloseExportReportRequest{
 		ExportHandle: int64(exportHandle),
 	}
-	_, err := client.GrpcClient.CloseExport(ctx, request)
+	_, err := client.GrpcClient.CloseExportReport(ctx, request)
 
 	return helper.ConvertGrpcError(err)
 }
@@ -2212,12 +2260,12 @@ func (client *Szengine) howEntityByEntityID(ctx context.Context, entityID int64,
 	return result, helper.ConvertGrpcError(err)
 }
 
-func (client *Szengine) preprocessRecord(ctx context.Context, recordDefinition string, flags int64) (string, error) {
-	request := &szpb.PreprocessRecordRequest{
+func (client *Szengine) getRecordPreview(ctx context.Context, recordDefinition string, flags int64) (string, error) {
+	request := &szpb.GetRecordPreviewRequest{
 		Flags:            flags,
 		RecordDefinition: recordDefinition,
 	}
-	response, err := client.GrpcClient.PreprocessRecord(ctx, request)
+	response, err := client.GrpcClient.GetRecordPreview(ctx, request)
 	result := response.GetResult()
 
 	return result, helper.ConvertGrpcError(err)