+++ szproduct/szproduct.go
@@ -57,6 +57,7 @@ func (client *Szproduct) GetLicense(ctx context.Context) (string, error) {
 		client.traceEntry(9)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(10, result, err, time.Since(entryTime)) }()
 	}
 
@@ -91,6 +92,7 @@ func (client *Szproduct) GetVersion(ctx context.Context) (string, error) {
 		client.traceEntry(11)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(12, result, err, time.Since(entryTime)) }()
 	}
 
@@ -123,6 +125,7 @@ func (client *Szproduct) Destroy(ctx context.Context) error {
 		client.traceEntry(3)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(4, err, time.Since(entryTime)) }()
 	}
 
@@ -176,6 +179,7 @@ func (client *Szproduct) Initialize(
 		client.traceEntry(13, instanceName, settings, verboseLogging)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(14, instanceName, settings, verboseLogging, err, time.Since(entryTime)) }()
 	}
 
@@ -207,6 +211,7 @@ func (client *Szproduct) RegisterObserver(ctx context.Context, observer observer
 		client.traceEntry(703, observer.GetObserverID(ctx))
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(704, observer.GetObserverID(ctx), err, time.Since(entryTime)) }()
 	}
 
@@ -242,6 +247,7 @@ func (client *Szproduct) SetLogLevel(ctx context.Context, logLevelName string) e
 		client.traceEntry(705, logLevelName)
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(706, logLevelName, err, time.Since(entryTime)) }()
 	}
 
@@ -290,6 +296,7 @@ func (client *Szproduct) UnregisterObserver(ctx context.Context, observer observ
 		client.traceEntry(707, observer.GetObserverID(ctx))
 
 		entryTime := time.Now()
+
 		defer func() { client.traceExit(708, observer.GetObserverID(ctx), err, time.Since(entryTime)) }()
 	}