+++ CHANGELOG.md
@@ -9,6 +9,18 @@ and this project adheres to [Semantic Versioning].
 
 -
 
+## [0.9.8] - 2025-07-02
+
+### Changed in 0.9.8
+
+- Change `SzConfig.AddDataSource` to `SzConfig.RegisterDataSource`
+- Change `SzConfig.DeleteDataSource` to `SzConfig.UnregisterDataSource`
+- Change `SzConfig.GetDataSources` to `SzConfig.GetDataSourceRegistry`
+- Change `SzDiagnostic.CheckDatastorePerformance` to `SzDiagnostic.CheckRepositoryPerformance`
+- Change `SzDiagnostic.GetDatastoreInfo` to `SzDiagnostic.GetRepositoryInfo`
+- Change `SzEngine.CloseExport` to `SzEngine.CloseExportReport`
+- Change `SzEngine.PreprocessRecord` to ``SzEngine.GetRecordPreview`
+
 ## [0.9.7] - 2025-06-17
 
 ### Changed in 0.9.7
@@ -29,7 +41,7 @@ and this project adheres to [Semantic Versioning].
 
 ## [0.9.5] - 2025-05-05
 
-### Changed in 0.9.5
+### Changed in 0.9.5a
 
 - Update dependencies